requirements.txt
Cython==0.28.2
kivy
pyserial
pygame
